generative/requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
sentence-transformers>=2.2.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 